added inserts strategy for upgrade to 2.5

+++ application/src/main/data/upgrade/2.4.3/schema_update_psql_ts.sql
@@ -36,6 +36,7 @@ BEGIN
     ALTER COLUMN key TYPE integer USING key::integer;
   ALTER TABLE ts_kv
       ADD CONSTRAINT ts_kv_pkey PRIMARY KEY (entity_id, key, ts);
+  CREATE TABLE IF NOT EXISTS ts_kv_indefinite PARTITION OF ts_kv DEFAULT;
 END;
 $$;
 
@@ -44,22 +45,40 @@ $$;
 CREATE OR REPLACE PROCEDURE create_new_ts_kv_latest_table() LANGUAGE plpgsql AS $$
 
 BEGIN
-  ALTER TABLE ts_kv_latest
-    RENAME TO ts_kv_latest_old;
-  ALTER TABLE ts_kv_latest_old
-     RENAME CONSTRAINT ts_kv_latest_pkey TO ts_kv_latest_pkey_old;
-  CREATE TABLE IF NOT EXISTS ts_kv_latest
-  (
-    LIKE ts_kv_latest_old
-  );
-  ALTER TABLE ts_kv_latest
-    DROP COLUMN entity_type;
-  ALTER TABLE ts_kv_latest
-    ALTER COLUMN entity_id TYPE uuid USING entity_id::uuid;
-  ALTER TABLE ts_kv_latest
-    ALTER COLUMN key TYPE integer USING key::integer;
-  ALTER TABLE ts_kv_latest
-    ADD CONSTRAINT ts_kv_latest_pkey PRIMARY KEY (entity_id, key);
+    IF NOT EXISTS(SELECT FROM pg_tables WHERE schemaname = 'public' AND tablename = 'ts_kv_latest_old') THEN
+          ALTER TABLE ts_kv_latest
+            RENAME TO ts_kv_latest_old;
+          ALTER TABLE ts_kv_latest_old
+             RENAME CONSTRAINT ts_kv_latest_pkey TO ts_kv_latest_pkey_old;
+          CREATE TABLE IF NOT EXISTS ts_kv_latest
+          (
+            LIKE ts_kv_latest_old
+          );
+          ALTER TABLE ts_kv_latest
+            DROP COLUMN entity_type;
+          ALTER TABLE ts_kv_latest
+            ALTER COLUMN entity_id TYPE uuid USING entity_id::uuid;
+          ALTER TABLE ts_kv_latest
+            ALTER COLUMN key TYPE integer USING key::integer;
+          ALTER TABLE ts_kv_latest
+            ADD CONSTRAINT ts_kv_latest_pkey PRIMARY KEY (entity_id, key);
+    ELSE
+        RAISE NOTICE 'ts_kv_latest_old table already exists!';
+        IF NOT EXISTS(SELECT FROM pg_tables WHERE schemaname = 'public' AND tablename = 'ts_kv_latest') THEN
+            CREATE TABLE IF NOT EXISTS ts_kv_latest
+            (
+                entity_id uuid   NOT NULL,
+                key       int    NOT NULL,
+                ts        bigint NOT NULL,
+                bool_v    boolean,
+                str_v     varchar(10000000),
+                long_v    bigint,
+                dbl_v     double precision,
+                json_v    json,
+                CONSTRAINT ts_kv_latest_pkey PRIMARY KEY (entity_id, key)
+            );
+        END IF;
+    END IF;
 END;
 $$;
 
@@ -218,4 +237,89 @@ BEGIN
 END;
 $$;
 
+-- call insert_into_ts_kv_cursor();
+
+CREATE OR REPLACE PROCEDURE insert_into_ts_kv_cursor() LANGUAGE plpgsql AS $$
+DECLARE
+    insert_size CONSTANT integer := 10000;
+    insert_counter       integer DEFAULT 0;
+    insert_record        RECORD;
+    insert_cursor CURSOR FOR SELECT to_uuid(entity_id)      	                                              AS entity_id,
+                                    ts_kv_records.key                                                         AS key,
+                                    ts_kv_records.ts                                                          AS ts,
+                                    ts_kv_records.bool_v                                                      AS bool_v,
+                                    ts_kv_records.str_v                                                       AS str_v,
+                                    ts_kv_records.long_v                                                      AS long_v,
+                                    ts_kv_records.dbl_v                                                       AS dbl_v
+                             FROM (SELECT entity_id                   AS entity_id,
+                                          key_id                      AS key,
+                                          ts,
+                                          bool_v,
+                                          str_v,
+                                          long_v,
+                                          dbl_v
+                                   FROM ts_kv_old
+                                            INNER JOIN ts_kv_dictionary ON (ts_kv_old.key = ts_kv_dictionary.key)) AS ts_kv_records;
+BEGIN
+    OPEN insert_cursor;
+    LOOP
+        insert_counter := insert_counter + 1;
+        FETCH insert_cursor INTO insert_record;
+        IF NOT FOUND THEN
+            RAISE NOTICE '% records have been inserted into the partitioned ts_kv!',insert_counter - 1;
+            EXIT;
+        END IF;
+        INSERT INTO ts_kv(entity_id, key, ts, bool_v, str_v, long_v, dbl_v)
+        VALUES (insert_record.entity_id, insert_record.key, insert_record.ts, insert_record.bool_v, insert_record.str_v,
+                insert_record.long_v, insert_record.dbl_v);
+        IF MOD(insert_counter, insert_size) = 0 THEN
+            RAISE NOTICE '% records have been inserted into the partitioned ts_kv!',insert_counter;
+        END IF;
+    END LOOP;
+    CLOSE insert_cursor;
+END;
+$$;
+
+-- call insert_into_ts_kv_latest_cursor();
+
+CREATE OR REPLACE PROCEDURE insert_into_ts_kv_latest_cursor() LANGUAGE plpgsql AS $$
+DECLARE
+    insert_size CONSTANT integer := 10000;
+    insert_counter       integer DEFAULT 0;
+    insert_record        RECORD;
+    insert_cursor CURSOR FOR SELECT to_uuid(entity_id)      	                                                     AS entity_id,
+                                    ts_kv_latest_records.key                                                         AS key,
+                                    ts_kv_latest_records.ts                                                          AS ts,
+                                    ts_kv_latest_records.bool_v                                                      AS bool_v,
+                                    ts_kv_latest_records.str_v                                                       AS str_v,
+                                    ts_kv_latest_records.long_v                                                      AS long_v,
+                                    ts_kv_latest_records.dbl_v                                                       AS dbl_v
+                             FROM (SELECT entity_id                   AS entity_id,
+                                          key_id                      AS key,
+                                          ts,
+                                          bool_v,
+                                          str_v,
+                                          long_v,
+                                          dbl_v
+                                   FROM ts_kv_latest_old
+                                            INNER JOIN ts_kv_dictionary ON (ts_kv_latest_old.key = ts_kv_dictionary.key)) AS ts_kv_latest_records;
+BEGIN
+    OPEN insert_cursor;
+    LOOP
+        insert_counter := insert_counter + 1;
+        FETCH insert_cursor INTO insert_record;
+        IF NOT FOUND THEN
+            RAISE NOTICE '% records have been inserted into the ts_kv_latest!',insert_counter - 1;
+            EXIT;
+        END IF;
+        INSERT INTO ts_kv_latest(entity_id, key, ts, bool_v, str_v, long_v, dbl_v)
+        VALUES (insert_record.entity_id, insert_record.key, insert_record.ts, insert_record.bool_v, insert_record.str_v,
+                insert_record.long_v, insert_record.dbl_v);
+        IF MOD(insert_counter, insert_size) = 0 THEN
+            RAISE NOTICE '% records have been inserted into the ts_kv_latest!',insert_counter;
+        END IF;
+    END LOOP;
+    CLOSE insert_cursor;
+END;
+$$;
